src/requirements.txt
pandas
pinotdb
streamlit
numpy
plotly-express
mysql-connector-python
msal_streamlit_authentication